updated routes in /chores to actually make sense. I don't think these are being used properly yet by the fetch requests. Update components so that the chore details and chore blocks are taking in a specific chore as a prop which will allow them to render a specific chore when the details button is clicked.

+++ backend/db/models/chore.js
@@ -4,7 +4,7 @@ module.exports = (sequelize, DataTypes) => {
     name: DataTypes.STRING,
     due_date: DataTypes.DATE,
     description: DataTypes.TEXT,
-    estimated_time: DataTypes.FLOAT,
+    estimated_time: DataTypes.INTEGER,
     isComplete: DataTypes.BOOLEAN,
     zone_id: DataTypes.INTEGER,
     user_id: DataTypes.INTEGER
